refactor: adapt windows

+++ uiviewer/_device.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 
 import abc
+import os
 import traceback
 import tempfile
-import os
-from typing import List, Dict, Union, Tuple, Optional
+from typing import List, Dict, Union, Tuple
 from functools import cached_property  # python3.8+
 
 from PIL import Image
@@ -56,14 +56,16 @@ class HarmonyDevice(DeviceMeta):
         return self.hdc.display_size()
 
     def take_screenshot(self) -> str:
-        png_base64 = None
-        with tempfile.NamedTemporaryFile(delete=True, suffix=".png") as f:
-            path = f.name
-            f.close()
+        temp_file = tempfile.NamedTemporaryFile(delete=False, suffix=".png")
+        try:
+            # adapt windows
+            temp_file.close()
+            path = temp_file.name
             self.hdc.screenshot(path)
-            png_base64 = file2base64(path)
-            os.remove(path)
-        return png_base64
+            return file2base64(path)
+        finally:
+            if os.path.exists(path):
+                os.remove(path)
 
     def dump_hierarchy(self) -> BaseHierarchy:
         packageName, pageName = self.hdc.current_app()
